PEP8: Renames IO module to tmt_io
