+++ Future/Tracker_Kinect.py
@@ -12,7 +12,7 @@ import time;
 import numpy as np;
 from math import sqrt;
 
-import TopMouseTracker.IO as IO;
+import TopMouseTracker.tmt_io as IO
 
 def getColorFrame(stream,resize):
