+++ _RunOffline.py
@@ -26,7 +26,7 @@ if os.getcwd() !=  _topMouseTrackerDir : #If the current working directory is no
 
 import TopMouseTracker.parameters as params
 import TopMouseTracker.utilities as utils
-import TopMouseTracker.IO as IO
+import TopMouseTracker.tmt_io as IO
 import TopMouseTracker._Tracker as tracker
 import TopMouseTracker.Analysis as analysis
 # Global parameters#
