+++ _Tracker.py
@@ -25,7 +25,8 @@ from PIL import Image;
 
 import TopMouseTracker.parameters as params
 import TopMouseTracker.utilities as utils
-import TopMouseTracker.IO as IO;
+import TopMouseTracker.tmt_io as IO
+
 
 class TopMouseTracker():
