+++ analysis.py
